+++ inductiva/api/__init__.py
@@ -1,5 +1,4 @@
 """Functionality to interact with the Web API."""
 from .methods import download_output, invoke_async_api, get_task_status
-from .methods import validate_api_key
+from .methods import get_api_config
 from .methods import get_client
-from .methods import compare_client_and_backend_versions
